Change polyfill selection change event name
Since the polyfill selection change event does not fire in the same
order as the non-polyfill one, and it fires in situations where the
native one does not (programatic DOM changes), I am renaming the event.
The event happens on the next animation frame, so I'm simply making it
the past tense.