Update delta, use invert in undo, fix typescript errors
We are updating our delta to the latest from quill-js, especially since they've added the invert method which we had just added ourselves before finding they had.

Because of issues with the output, we are including the source rather than using it from npm. Downstream projects using webpack were importing modules before they were defined. I assume this can be fixed with `.tsconfig` files settings in Delta, but research needs to be done. This needs to be looked into further to find a proper solution where we don't need to copy the files into our project.